Appraise Rails 7.1
