+++ Appraisals
@@ -2,12 +2,6 @@
 
 case RUBY_VERSION
 when '2.7.7', '3.1.3', '3.2.1'
-  appraise "ruby-#{RUBY_VERSION}_activerecord60" do
-    source 'https://rubygems.org' do
-      gem 'activerecord', '~> 6.0.0'
-    end
-  end
-
   appraise "ruby-#{RUBY_VERSION}_activerecord61" do
     source 'https://rubygems.org' do
       gem 'activerecord', '~> 6.1.0'
@@ -17,6 +11,12 @@ when '2.7.7', '3.1.3', '3.2.1'
   appraise "ruby-#{RUBY_VERSION}_activerecord70" do
     source 'https://rubygems.org' do
       gem 'activerecord', '~> 7.0.0'
+    end
+  end
+
+  appraise "ruby-#{RUBY_VERSION}_activerecord71" do
+    source 'https://rubygems.org' do
+      gem 'activerecord', '~> 7.1.0'
     end
   end
 else
