+++ Gemfile
@@ -3,10 +3,10 @@
 source 'https://rubygems.org' # global source
 
 source 'https://rubygems.org' do
-  gem 'activesupport', '>= 6', '< 7.1'
+  gem 'activesupport', '>= 6.1', '< 7.2'
   gem 'appraisal', '>= 2.4', '< 3'
   gem 'bundler', '>= 2.2', '< 3'
-  gem 'minitest', '>= 5.19', '< 6'
+  gem 'minitest', '>= 5.20', '< 6'
   gem 'minitest-reporters', '>= 1.6', '< 2'
   gem 'mysql2', '>= 0.5.5', '< 1'
   gem 'rake', '>= 13', '< 14'
